separate file for class and rename class
